Added SFXManager to play sound effects
Added sound effects manager as a singelton. Sounds can be imported into the manager and call from any script with SFXManager.sfxinstance.Audio.PlayOneShot(SFXManager.sfxinstance.SoundEffectName); where SoundEffectName is replaced by the name of the variable that is defined in the SFXManager (add a new variable in there if you need a new sound.